+++ packages/modal/src/components/modal/ModalFooter.tsx
@@ -1,18 +1,24 @@
-import { BoxProps, Column } from "@stenajs-webui/core";
+import { Column } from "@stenajs-webui/core";
 import * as React from "react";
 
-interface ModalFooterProps extends Omit<BoxProps, "indent" | "spacing"> {
-  sticky?: boolean;
+interface ModalFooterProps {
+  disableSticky?: boolean;
 }
 
-export const ModalFooter: React.FC<ModalFooterProps> = (props) => (
+export const ModalFooter: React.FC<ModalFooterProps> = ({
+  disableSticky,
+  children,
+}) => (
   <Column
     indent={2}
     spacing={2}
     background={"var(--swui-modal-content-bg-color)"}
-    position={props.sticky ? undefined : "sticky"}
-    bottom={props.sticky ? undefined : "calc(-1 * var(--swui-modal-padding))"}
-    shadow={props.sticky ? undefined : "var(--swui-sticky-modal-footer-shadow)"}
-    {...props}
-  />
+    position={disableSticky ? undefined : "sticky"}
+    bottom={disableSticky ? undefined : "calc(-1 * var(--swui-modal-padding))"}
+    shadow={
+      disableSticky ? undefined : "var(--swui-sticky-modal-footer-shadow)"
+    }
+  >
+    {children}
+  </Column>
 );